+++ R/utils.R
@@ -27,11 +27,11 @@ unpack_assignments <- function(source, parsed,
   wrapr::stop_if_dot_args(substitute(list(...)),
                           "rquery:::unpack_assignments")
   n <- length(parsed)
+  assignments <- character(n)
   if(n<=0) {
-    stop("must generate at least 1 column")
+    return(assignments)
   }
   nms <- character(n)
-  assignments <- character(n)
   uses <- vector(n, mode='list')
   for(i in seq_len(n)) {
     si <- parsed[[i]]